requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
flask-cors==3.0.10
python-dotenv==1.0.0
elevenlabs==1.9.0
pydantic==1.10.9
moviepy==1.0.3
youtube_dl==2021.12.17
python-decouple==3.8